tests/wpt/tests/css/css-display/display-first-letter-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Display: first-letter pseudo-element</title>
<link rel="help" href="https://www.w3.org/TR/css-display-3/#placement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #t1::first-letter { float: left; display: flex; font-size: 30px }
  #t2::first-letter { float: left; font-size: 30px }
  #t3::first-letter { display: flex; font-size: 30px }
  #t4::first-letter { font-size: 30px }
</style>
<div id="t1">First letter is float and flex.</div>
<div id="t2">First letter is float but not flex.</div>
<div id="t3">First letter is flex but not float.</div>
<div id="t4">First letter not float or flex.</div>
<script>
  function getFirstLetterDisplayFor(id) {
    return window.getComputedStyle(document.getElementById(id), "::first-letter").display;
  }
  test(function() {
    assert_equals(getFirstLetterDisplayFor("t1"), "block");
    assert_equals(getFirstLetterDisplayFor("t2"), "block");
    assert_equals(getFirstLetterDisplayFor("t3"), "inline");
    assert_equals(getFirstLetterDisplayFor("t4"), "inline");
  }, "display of first-letter");
</script>
